Fix history bug
